change data mapper structure
