--- app/src/main/java/com/example/movieapp/data/mapper/DataMapper.kt
@@ -1,7 +1,0 @@
-package com.example.movieapp.data.mapper
-
-interface DataMapper<in DTO, DOMAIN, ENTITY> {
-    fun dtoToDomain(dto: DTO): List<DOMAIN>
-    fun fromEntity(entity: List<ENTITY>): List<DOMAIN>
-    fun toEntity(favoriteMovie: DOMAIN): ENTITY
-}
