+++ app/src/main/java/com/example/movieapp/data/mapper/DtoMapper.kt
@@ -1,0 +1,5 @@
+package com.example.movieapp.data.mapper
+
+interface DtoMapper<in DTO, out DOMAIN> {
+    fun dtoToDomain(dto: DTO): List<DOMAIN>
+}
