+++ app/src/main/java/com/example/movieapp/data/mapper/EntityMapper.kt
@@ -1,0 +1,6 @@
+package com.example.movieapp.data.mapper
+
+interface EntityMapper<ENTITY, DOMAIN> {
+    fun fromEntity(entity: List<ENTITY>): List<DOMAIN>
+    fun toEntity(domain: DOMAIN): ENTITY
+}
